+++ FitCloudKit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## V1.3.2-beta.15 build20251103001 (2025-11-03)
+
+- Modified the SDK initialization method to return a boolean indicating success instead of the FitCloudKit object.
+- Introduce new sdk initialize option `logLevel` to set the minimum log level for the SDK.
+- Introduce new sdk initialize option `includeTimestampInLogs` to indicate whether to include the timestamp in the log messages.
+- Renamed the SDK log message callback Method from `- (void)OnLogMessage:(NSString *)message level:(FITCLOUDKITLOGLEVEL)level;` to `- (void)onLogMessage:(NSString *)message level:(FITCLOUDKITLOGLEVEL)level subsystem:(NSString *)subsystem category:(NSString *)category;`.
+- Additional SDK stability improvements and performance optimizations.
+
 ## V1.3.2-beta.14 build20251031001 (2025-10-31)
 
 - Updated the task-sending API to allow setting `totalTaskCount` and `totalCoinsEarned` even when the task list is empty or `nil` .